+++ CHANGELOG.md
@@ -9,5 +9,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ClimateDT workflow modifications:
 
 AQUA-diagnostics complete list:
+- Porting of the AQUA diagnostics documentation (#18)
 - Added GitHub Issues and PR templates (#1)
 - CHANGELOG, LICENSE and README files added (#2)